requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
tensorflow-lite==2.16.1
pillow==10.1.0
numpy==1.24.3
gunicorn==20.1.0 